Fix errors in the TwoApps showcase example (#1078)
* Use the `Msg::SetOpposite` in `two_apps` example

b4d5227e625640ee2aca548ee06499dfe90c8acf changed the name of
`Msg::SetScope` to `Msg::SetOpposite` but the usages in the `main`
functions of the `two_apps` examples were not updated.

* Use new `TwoModels` struct as the `TwoApps` scene

`TwoModels` is a `Component` that does not respond to changes but which
creates two instances of the `Model` and mounts them as separate
`yew::App<Model>` instances in the body. This is done to mimic the
behavior of `two_apps/src/main.rs` which uses a static `index.html` file
containing the `div.first-app` and `div.second-app` elements.
`TwoModels::mounted` is responsible for replacing the `<div>` nodes it
creates during the initial render with the `Model` components.